Version 0.1 Basic Flask Application
Current Version 0.1:
I have developed some basic flask templates along with a CSS stylesheet. The flask app contains some basic SQLite code that allows for posts to be uploaded. The login page has no functionality. The signup page has no functionality. The main page is the first view a user has, it has some basic functionality, it presents user created posts in a list format. 

Future:
Login and Signup pages need to connect to the database. When a user signs up, validation needs to be done to check if that user already exists in the database. It then needs to create an account for them. And allow them to sign in. 

The forgot password needs to check if the email/user exists and needs to send a password reset email. 

The profile needs to be updated to contain the ability to change user settings such as username, password, email.

And finally, the main needs to be updated with AJAX code that allows the user generated titles and descriptions to be their own individual post. In addition, editing, liking, deleting and commenting on posts needs to be implemented.  

Highest Priority:
Ensuring the signup and sign function. I will recommend referring to Flask Web Development by Grinberg, Miguel chapter 2 part 2 because it contains an incredibly useful example of how to implement a flask app.